More extraction.
