+++ cryptomator/src/main/java/ch/cyberduck/core/cryptomator/features/CryptoCopyFeature.java
@@ -19,7 +19,7 @@ import ch.cyberduck.core.ConnectionCallback;
 import ch.cyberduck.core.Path;
 import ch.cyberduck.core.PathAttributes;
 import ch.cyberduck.core.Session;
-import ch.cyberduck.core.cryptomator.CryptoVaultInterface;
+import ch.cyberduck.core.cryptomator.AbstractVault;
 import ch.cyberduck.core.cryptomator.random.RandomNonceGenerator;
 import ch.cyberduck.core.cryptomator.random.RotatingNonceGenerator;
 import ch.cyberduck.core.exception.BackgroundException;
@@ -37,11 +37,11 @@ public class CryptoCopyFeature implements Copy {
 
     private final Session<?> session;
     private final Copy proxy;
-    private final CryptoVaultInterface vault;
+    private final AbstractVault vault;
 
     private Session<?> target;
 
-    public CryptoCopyFeature(final Session<?> session, final Copy proxy, final CryptoVaultInterface vault) {
+    public CryptoCopyFeature(final Session<?> session, final Copy proxy, final AbstractVault vault) {
         this.session = session;
         this.target = session;
         this.proxy = proxy;
