+++ cryptomator/src/main/java/ch/cyberduck/core/cryptomator/features/CryptoDirectoryV6Feature.java
@@ -19,8 +19,8 @@ import ch.cyberduck.core.Path;
 import ch.cyberduck.core.RandomStringService;
 import ch.cyberduck.core.Session;
 import ch.cyberduck.core.UUIDRandomStringService;
+import ch.cyberduck.core.cryptomator.AbstractVault;
 import ch.cyberduck.core.cryptomator.ContentWriter;
-import ch.cyberduck.core.cryptomator.CryptoVaultInterface;
 import ch.cyberduck.core.cryptomator.random.RandomNonceGenerator;
 import ch.cyberduck.core.exception.BackgroundException;
 import ch.cyberduck.core.features.Directory;
@@ -40,11 +40,11 @@ public class CryptoDirectoryV6Feature<Reply> implements Directory<Reply> {
     private final Session<?> session;
     private final Write<Reply> writer;
     private final Directory<Reply> delegate;
-    private final CryptoVaultInterface vault;
+    private final AbstractVault vault;
     private final RandomStringService random = new UUIDRandomStringService();
 
     public CryptoDirectoryV6Feature(final Session<?> session, final Directory<Reply> delegate,
-                                    final Write<Reply> writer, final CryptoVaultInterface cryptomator) {
+                                    final Write<Reply> writer, final AbstractVault cryptomator) {
         this.session = session;
         this.writer = writer;
         this.delegate = delegate;
